+++ pyoupyou/settings/dev.py
@@ -51,3 +51,4 @@ MAIL_HR = "hr@enioka.com"
 MAIL_FROM = "pyoupyou@enioka.com"
 
 SEEKUBE_SOURCE_ID = 5
+SECRET_ANON_SALT = "dev-salt"